Removed unnecessary div or imports from Profile components
