+++ frontend/src/components/Profile/ChangeAvatarBox.tsx
@@ -1,0 +1,56 @@
+import { Box, Button } from "@mui/material";
+
+export function ChangeAvatarBox({handleAvatarSelected}) {
+  return (
+    <Box
+      sx={{
+        border: "1px solid black",
+        borderRadius: "5px",
+        margin: "20px",
+        width: "50%",
+        display: "flex",
+        justifyContent: "center", // Center horizontally
+        alignItems: "center", // Center vertically
+      }}
+    >
+      <input
+        type="file" // Indicated that we will select a type file
+        accept="image/*" // It only indicates that I accept images
+        onChange={handleAvatarSelected}
+        style={{ display: "none" }} // Hide the input element totally
+        id="avatar-input" // Creates a specific id for the input
+      />
+      <label htmlFor="avatar-input">
+        <Button variant="contained" size="large" component="span">
+          Change Avatar
+        </Button>
+      </label>
+    </Box>
+  );
+}
+
+//Content to put inside Profile, if no way to update profile. 
+{/* <Box
+sx={{
+  border: "1px solid black",
+  borderRadius: "5px",
+  margin: "20px",
+  width: "50%",
+  display: "flex",
+  justifyContent: "center", // Center horizontally
+  alignItems: "center", // Center vertically
+}}
+>
+<input
+  type="file" // Indicated that we will select a type file
+  accept="image/*" // It only indicates that I accept images
+  onChange={handleAvatarSelected}
+  style={{ display: "none" }} // Hide the input element totally
+  id="avatar-input" // Creates a specific id for the input
+/>
+<label htmlFor="avatar-input">
+  <Button variant="contained" size="large" component="span">
+	Change Avatar
+  </Button>
+</label>
+</Box> */}
