+++ frontend/src/pages/Profile.tsx
@@ -9,7 +9,8 @@ import { ContentBox } from "../components/Profile/ContentBox";
 import { ContainerGrid } from "../components/Profile/ContainerGrid";
 import { RightSideGrid } from "../components/Profile/RightSideGrid";
 import { LeftSideGrid } from "../components/Profile/LeftSideGrid";
-
+import { ChangeAvatarBox } from "../components/Profile/ChangeAvatarBox";
+import { ChangeNameBox } from "../components/Profile/ChangeNameBox";
 
 export function Profile() {
   const [user, setUser] = useState<User>();
@@ -34,9 +35,6 @@ export function Profile() {
     fetchUsersData();
   }, []);
 
-  function handleUsername() {
-	console.log("Click Button");
-  }
 
   const handleAvatarSelected = (event) => {
     const selectedFile = event.target.files[0];
@@ -69,53 +67,9 @@ export function Profile() {
 	<ContainerGrid>
       <LeftSideGrid>
 		<NameBox />           
-		<AvatarBox/>
-        <Box
-          sx={{
-            border: "1px solid black",
-            borderRadius: "5px",
-            margin: "20px",
-            width: "50%",
-            display: "flex",
-            justifyContent: "center", // Center horizontally
-            alignItems: "center", // Center vertically
-          }}
-        >
-          <input
-            type="file" // Indicated that we will select a type file
-            accept="image/*" // It only indicates that I accept images
-            onChange={handleAvatarSelected}
-            style={{ display: "none" }} // Hide the input element totally
-            id="avatar-input" // Creates a specific id for the input
-          />
-          <label htmlFor="avatar-input">
-            <Button variant="contained" size="large" component="span">
-              Change Avatar
-            </Button>
-          </label>
-        </Box>
-        <Box
-          sx={{
-            border: "1px solid black",
-            borderRadius: "5px",
-            margin: "20px",
-            width: "50%",
-            display: "flex",
-            justifyContent: "center", // Center horizontally
-            alignItems: "center", // Center vertically
-          }}
-        >
-          <input
-            type="text" // Indicated that we will select a type file
-            style={{ display: "none" }} // Hide the input element totally
-            id="username-input" // Creates a specific id for the input
-          />
-          <label htmlFor="username-input">
-            <Button variant="contained" size="large" component="span" onClick={handleUsername}>
-              Change Username
-            </Button>
-          </label>
-        </Box>
+		<AvatarBox user={user}/>
+		<ChangeAvatarBox onChange={handleAvatarSelected}/>
+		<ChangeNameBox/>
 	  </LeftSideGrid>
 	 <RightSideGrid>
 		<ContentBox content="ABOUT"/>
